Refactor: move files & change routing path
