+++ src/pages/Auth/FindPassword/FindPassword.tsx
@@ -1,0 +1,11 @@
+import Header from "@/components/Auth/Header/Header";
+
+function FindPassword() {
+  return (
+    <div>
+      <Header />
+    </div>
+  );
+}
+
+export default FindPassword;
